Move implementation of Server out of the header.
